+++ docs/source/Changelog.rst
@@ -84,6 +84,11 @@ Added
       fields (#739)
    -  New ``sort_style`` named ``field`` and ``sort_field`` option to
       sort by one or more arbitrary fields.
+   -  New ``csv`` options: (to match KiCad internal BoM functionality)
+
+      -  ``string_delimiter`` to configure the quote character
+      -  ``keep_line_breaks`` to remove line breaks in fields
+      -  ``keep_tabs`` to remove tabs in fields
 
 -  Draw Fancy Stackup: